# contracts/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "ErrorResponse" : {
        "description" : "Error response",
        "type" : "object",
        "required" : [ "code" ],
        "properties" : {
          "code" : {
            "description" : "Error code",
            "type" : "string",
            "$ref" : "#/components/schemas/ErrorCode"
          },
          "errorDetails" : {
            "description" : "Additional details",
            "type" : "object",
            "$ref" : "#/components/schemas/ErrorDetails"
          }
        }
      },
...
      "MediaDto" : {
        "type" : "object",
        "required" : [ "markdown", "assets" ],
        "properties" : {
          "markdown" : {
            "type" : "string",
            "description" : "Markdown"
          },
          "assets" : {
            "description" : "Assets",
            "type" : "object",
            "$ref" : "#/components/schemas/AssetsDto"
          }
        }
      },
...
      "MinRole" : {
        "type" : "string",
        "enum" : [ "NOT_MEMBER", "MEMBER", "ORGANIZER", "ADMIN" ]
      },
...
      "TeamListResponse" : {
        "description" : "Paginated team list response",
        "type" : "object",
        "required" : [ "teams", "total", "page", "size" ],
        "properties" : {
          "teams" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TeamDetailDto"
            },
            "description" : "List of teams"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Total number of teams"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Current page number"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Page size"
          }
        }
      },
...
      "TeamRequest" : {
        "description" : "Team creation request",
        "type" : "object",
        "required" : [ "name", "media", "visibility", "enableTrips", "enableAds" ],
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "Team name",
            "examples" : [ "Awesome Cycling Team" ],
            "pattern" : "\\S",
            "minLength" : 1,
            "maxLength" : 200
          },
          "media" : {
            "description" : "Media",
            "type" : "object",
            "$ref" : "#/components/schemas/MediaDto"
          },
          "visibility" : {
            "description" : "Team visibility",
            "type" : "string",
            "examples" : [ "PUBLIC" ],
            "$ref" : "#/components/schemas/Visibility"
          },
          "enableTrips" : {
            "type" : "boolean",
            "description" : "Trips enabled for team",
            "examples" : [ true ]
          },
          "enableAds" : {
            "type" : "boolean",
            "description" : "Ads enabled for team",
            "examples" : [ true ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "openIdConnectUrl" : "http://localhost:8180/realms/quarkus/.well-known/openid-configuration",
        "description" : "Authentication"
      }
    }
  },
...
  "paths" : {
...
    "/api/teams" : {
      "get" : {
        "summary" : "List public teams",
        "description" : "Get a paginated list of public teams with optional search",
        "operationId" : "listTeams",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Minimum role in team",
          "name" : "minRole",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/MinRole"
          }
        }, {
          "description" : "Page number (0-indexed)",
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "description" : "Search query to filter teams by name",
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Page size",
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Teams retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamListResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create team",
        "description" : "Create a new team. The current user will be set as the team owner.",
        "operationId" : "createTeam",
        "tags" : [ "Teams" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Team created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/teams/{slug}" : {
      "put" : {
        "summary" : "Update team",
        "description" : "Update team information. Requires ADMIN role.",
        "operationId" : "updateTeam",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Team updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      },
      "get" : {
        "summary" : "Get team by slug",
        "description" : "Get detailed team information by URL slug",
        "operationId" : "getTeam",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Team retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Team is private and user is not a member",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete team",
        "description" : "Soft delete a team. Requires OWNER role.",
        "operationId" : "deleteTeam",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Team deleted successfully"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not the team owner",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
    "/api/teams/{slug}/posts" : {
      "get" : {
        "summary" : "List posts",
        "description" : "Get paginated list of posts for a team with optional filtering",
        "operationId" : "listPosts",
        "tags" : [ "Posts" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Start date filter (ISO format)",
          "name" : "from",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Page number",
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "description" : "Search by name/description",
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Page size",
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "description" : "End date filter (ISO format)",
          "name" : "to",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Posts retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostListResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create post",
        "description" : "Create a new post with optional groups",
        "operationId" : "createPost",
        "tags" : [ "Posts" ],
        "parameters" : [ {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PostRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Post created successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team member",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
    "/api/teams/{slug}/posts/{postSlug}" : {
      "put" : {
        "summary" : "Update post",
        "description" : "Update post information. Requires organizer permissions.",
        "operationId" : "updatePost",
        "tags" : [ "Posts" ],
        "parameters" : [ {
          "description" : "Post URL slug",
          "name" : "postSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PostRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Post updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not authorized to update this post",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or post not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      },
      "get" : {
        "summary" : "Get post details",
        "description" : "Get detailed post information including groups",
        "operationId" : "getPost",
        "tags" : [ "Posts" ],
        "parameters" : [ {
          "description" : "Post URL slug",
          "name" : "postSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Post retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or post not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete post",
        "description" : "Soft delete a post. Requires organizer permissions.",
        "operationId" : "deletePost",
        "tags" : [ "Posts" ],
        "parameters" : [ {
          "description" : "Post URL slug",
          "name" : "postSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Post deleted successfully"
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not authorized to delete this post",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team or post not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
    "/api/teams/{slug}/posts/{postSlug}/slug" : {
      "patch" : {
        "summary" : "Change post slug",
        "description" : "Change post URL slug. Requires organizer permissions.",
        "operationId" : "changePostSlug",
        "tags" : [ "Posts" ],
        "parameters" : [ {
          "description" : "Current post URL slug",
          "name" : "postSlug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SlugChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Slug changed successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PostDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid slug format",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Slug already in use",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
    "/api/teams/{slug}/slug" : {
      "patch" : {
        "summary" : "Change team slug",
        "description" : "Change team URL slug. Requires ADMIN role.",
        "operationId" : "changeTeamSlug",
        "tags" : [ "Teams" ],
        "parameters" : [ {
          "description" : "Current team URL slug",
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SlugChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Slug changed successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamDetailDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid slug format",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "User is not a team admin",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Slug already in use",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "user" ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "Tribly API",
    "version" : "1.0.0",
    "description" : "API for Tribly Cycling Team Management Platform"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}